+++ lib/src/inverted_index/async_index.dart
@@ -6,48 +6,13 @@
 
 import 'package:text_indexing/text_indexing.dart';
 
-/// The [AsyncCallbackIndex] is a [InvertedIndex] implementation class
-/// that uses asynchronous callbacks to perform read and write operations on
-/// [Dictionary] and [Postings] repositories:
-/// - [phraseLength] is the maximum length of phrases in the index vocabulary.
-///   The minimum phrase length is 1. If phrase length is greater than 1, the
-///   index vocabulary also contains phrases up to [phraseLength] long,
-///   concatenated from consecutive terms. The index size is increased by a
-///   factor of [phraseLength];
-/// - [analyzer] is the [ITextAnalyzer] used to tokenize text for the index;
-/// - [vocabularyLength] is the number of unique terms in the corpus;
-/// - [zones] is a hashmap of zone names to their relative weight in the index;
-/// - [k] is the length of k-gram entries in the k-gram index;
-/// - [getDictionary] Asynchronously retrieves a [Dictionary] for a collection
-///   of [Term]s from a [Dictionary] repository;
-/// - [upsertDictionary ] inserts entries into a [Dictionary] repository,
-///   overwriting any existing entries;
-/// - [getKGramIndex] Asynchronously retrieves a [KGramIndex] for a collection
-///   of [KGram]s from a [KGramIndex] repository;
-/// - [upsertKGramIndex ] inserts entries into a [KGramIndex] repository,
-///   overwriting any existing entries;
-/// - [getPostings] asynchronously retrieves [Postings] for a collection
-///   of [Term]s from a [Postings] repository;
-/// - [upsertPostings] inserts entries into a [Postings] repository,
-///   overwriting any existing entries;
-/// - [getFtdPostings] return a [FtdPostings] for a collection of [Term]s from
-///   the [Postings], optionally filtered by minimum term frequency; and
-/// - [getIdFtIndex] returns a [IdFtIndex] for a collection of [Term]s from
-///   the [Dictionary];
-/// - [dictionaryLoader] asynchronously retrieves a [Dictionary] for a vocabulary
-///   from a index repository;
-/// - [dictionaryLengthLoader] asynchronously retrieves the number of terms in
-///   the vocabulary (N);
-/// - [dictionaryUpdater] is callback that passes a [Dictionary] subset
-///    for persisting to a index repository;
-/// - [kGramIndexLoader] asynchronously retrieves a [KGramIndex] for a vocabulary
-///   from a index repository;
-/// - [kGramIndexUpdater] is callback that passes a [KGramIndex] subset
-///    for persisting to a index repository;
-/// - [postingsLoader] asynchronously retrieves a [Postings] for a vocabulary
-///   from a index repository; and
-/// - [postingsUpdater] passes a [Postings] subset for persisting to a
-///   index repository.
+/// The [AsyncCallbackIndex] is a [InvertedIndex] implementation class that
+/// mixes in [InvertedIndexMixin] and [AsyncCallbackIndexMixin].
+///
+/// The InMemoryIndex is intended for working with a larger corpus and an
+/// asynchronous index repository in persisted storage.  It uses asynchronous
+/// callbacks to perform read and write operations on [Dictionary], [KGramIndex]
+/// and [Postings] repositories.
 class AsyncCallbackIndex
     with InvertedIndexMixin, AsyncCallbackIndexMixin
     implements InvertedIndex {
@@ -81,23 +46,20 @@ class AsyncCallbackIndex
   final PostingsUpdater postingsUpdater;
 
   /// Instantiates a [AsyncCallbackIndex] instance:
-  /// - [phraseLength] is the maximum length of phrases in the index vocabulary.
-  ///   The minimum phrase length is 1. If phrase length is greater than 1, the
-  ///   index vocabulary also contains phrases up to [phraseLength] long,
-  ///   concatenated from consecutive terms. The index size is increased by a
-  ///   factor of [phraseLength];
   /// - [analyzer] is the [ITextAnalyzer] used to tokenize text for the index;
-  /// - [vocabularyLength] is the number of unique terms in the corpus;
-  /// - [zones] is a hashmap of zone names to their relative weight in the index;
   /// - [k] is the length of k-gram entries in the k-gram index;
+  /// - [zones] is a hashmap of zone names to their relative weight in the
+  ///   index;
+  /// - [phraseLength] is the maximum length of phrases in the index vocabulary
+  ///   and must be greater than 0;
   /// - [dictionaryLengthLoader] asynchronously retrieves the number of terms
   ///   in the vocabulary (N);
-  /// - [dictionaryLoader] asynchronously retrieves a [Dictionary] for a vocabulary
-  ///   from a index repository;
+  /// - [dictionaryLoader] asynchronously retrieves a [Dictionary] for a
+  ///   vocabulary from a index repository;
   /// - [dictionaryUpdater] is callback that passes a [Dictionary] subset
   ///    for persisting to a index repository;
-  /// - [kGramIndexLoader] asynchronously retrieves a [KGramIndex] for a vocabulary
-  ///   from a index repository;
+  /// - [kGramIndexLoader] asynchronously retrieves a [KGramIndex] for a
+  ///   vocabulary from a index repository;
   /// - [kGramIndexUpdater] is callback that passes a [KGramIndex] subset
   ///    for persisting to a index repository;
   /// - [postingsLoader] asynchronously retrieves a [Postings] for a vocabulary