+++ lib/src/inverted_index/cached_index.dart
@@ -6,39 +6,14 @@ import 'dart:async';
 
 import 'package:text_indexing/text_indexing.dart';
 
-/// An implementation of the [InvertedIndex] interface:
-/// - [phraseLength] is the maximum length of phrases in the index vocabulary.
-///   The minimum phrase length is 1. If phrase length is greater than 1, the
-///   index vocabulary also contains phrases up to [phraseLength] long,
-///   concatenated from consecutive terms. The index size is increased by a
-///   factor of [phraseLength];
-/// - [analyzer] is the [ITextAnalyzer] used to tokenize text for the index;
-/// - [vocabularyLength] is the number of unique terms in the corpus;
-/// - [zones] is a hashmap of zone names to their relative weight in the index;
-/// - [k] is the length of k-gram entries in the k-gram index;
-/// - [cacheLimit] is the maximum number of entries in any of the caches;
-/// - [getDictionary] retrieves a [Dictionary] for a collection of [Term]s from
-///   the in-memory [dictionaryCache] hashmap;
-/// - [upsertDictionary ] inserts entries into the in-memory [dictionaryCache] hashmap,
-///   overwriting any existing entries;
-/// - [getKGramIndex] Asynchronously retrieves a [KGramIndex] for a collection
-///   of [KGram]s from a [KGramIndex] repository;
-/// - [upsertKGramIndex ] inserts entries into a [KGramIndex] repository,
-///   overwriting any existing entries;
-/// - [getPostings] retrieves [Postings] for a collection of [Term]s from the
-///   in-memory [postingsCache] hashmap;
-/// - [upsertPostings] inserts entries into the in-memory [postingsCache] hashmap,
-///   overwriting any existing entries;
-/// - [getFtdPostings] return a [FtdPostings] for a collection of [Term]s from
-/// the [Postings], optionally filtered by minimum term frequency; and
-/// - [getIdFtIndex] returns a [IdFtIndex] for a collection of [Term]s from
-/// the [Dictionary].
-/// - [dictionaryCache] is the in-memory term dictionaryCache for the indexer. Pass a
-///   [dictionaryCache] instance at instantiation, otherwise an empty [Dictionary]
-///   will be initialized; and
-/// - [postingsCache] is the in-memory postingsCache hashmap for the indexer. Pass a
-///   [postingsCache] instance at instantiation, otherwise an empty [Postings]
-///   will be initialized.
+/// The [CachedIndex] is a [InvertedIndex] implementation class that mixes in
+/// [InvertedIndexMixin] and [CachedIndexMixin].
+///
+/// The CachedIndex is intended for working with a larger corpus with an
+/// asynchronous index repository in persisted storage.  It uses asynchronous
+/// callbacks to perform read and write operations on [Dictionary], [KGramIndex]
+/// and [Postings] repositories, but keeps a cache of the most popular terms
+/// and k-grams in memory for faster indexing and searching.
 class CachedIndex
     with CachedIndexMixin, InvertedIndexMixin
     implements InvertedIndex {
@@ -90,20 +65,23 @@ class CachedIndex
 
   /// Instantiates a [InMemoryIndex] instance:
   /// - [analyzer] is the [ITextAnalyzer] used to tokenize text for the index;
+  /// - [k] is the length of k-gram entries in the k-gram index;
+  /// - [zones] is a hashmap of zone names to their relative weight in the
+  ///   index;
+  /// - [phraseLength] is the maximum length of phrases in the index vocabulary
+  ///   and must be greater than 0;
   /// - [cacheLimit] is the maximum number of entries in any of the caches;
-  /// - [zones] is a hashmap of zone names to their relative weight in the index;
   /// - [dictionaryCache] is the in-memory term dictionaryCache for the indexer. Pass a
   ///   [dictionaryCache] instance at instantiation, otherwise an empty [Dictionary]
   ///   will be initialized;
-  /// - [phraseLength] is the maximum length of phrases in the index vocabulary.
   /// - [dictionaryLengthLoader] asynchronously retrieves the number of terms
   ///   in the vocabulary (N);
-  /// - [dictionaryLoader] asynchronously retrieves a [Dictionary] for a vocabulary
-  ///   from a index repository;
+  /// - [dictionaryLoader] asynchronously retrieves a [Dictionary] for a
+  ///   vocabulary from a index repository;
   /// - [dictionaryUpdater] is callback that passes a [Dictionary] subset
   ///    for persisting to a index repository;
-  /// - [kGramIndexLoader] asynchronously retrieves a [KGramIndex] for a vocabulary
-  ///   from a index repository;
+  /// - [kGramIndexLoader] asynchronously retrieves a [KGramIndex] for a
+  ///   vocabulary from a index repository;
   /// - [kGramIndexUpdater] is callback that passes a [KGramIndex] subset
   ///    for persisting to a index repository;
   /// - [postingsLoader] asynchronously retrieves a [Postings] for a vocabulary